+++ frontend/src/pages/LandSurveyForm/LandSurveyCreate.tsx
@@ -358,7 +358,7 @@ const LandSurveyCreate = () => {
                         type="number"
                         id="surface"
                         value={surface}
-                        onChange={(e) => setSurface(parseInt(e.target.value))}
+                        onChange={(e) => setSurface(parseFloat(e.target.value))}
                         required
                     />
                 </div>
